embed widget link for qgis-server and geoserver backend
embed widget link and add copy button

fix #152
