+++ geonode/maps/urls.py
@@ -39,17 +39,13 @@ if check_ogc_backend(geoserver.BACKEND_PACKAGE):
     map_embed = 'map_embed'
     # TODO qlr for geoserver
     map_download_qlr = 'map_download_qlr'
-    # TODO no implementation for geoserver
-    map_embed_widget = 'map_embed_widget'
 
 elif check_ogc_backend(qgis_server.BACKEND_PACKAGE):
     new_map_view = MapCreateView.as_view()
     existing_map_view = MapDetailView.as_view()
     map_embed = MapEmbedView.as_view()
-    from geonode.maps.qgis_server_views import map_download_qlr, \
-        map_embed_widget
+    from geonode.maps.qgis_server_views import map_download_qlr
     map_download_qlr = map_download_qlr
-    map_embed_widget = map_embed_widget
 
 urlpatterns = patterns(
     'geonode.maps.views',
@@ -75,7 +71,7 @@ urlpatterns = patterns(
     url(r'^(?P<mapid>[^/]+)/metadata$', 'map_metadata', name='map_metadata'),
     url(r'^(?P<mapid>[^/]+)/metadata_advanced$', 'map_metadata_advanced', name='map_metadata_advanced'),
     url(r'^(?P<mapid>[^/]+)/embed$', map_embed, name='map_embed'),
-    url(r'^(?P<mapid>[^/]+)/embed_widget$', map_embed_widget, name='map_embed_widget'),
+    url(r'^(?P<mapid>[^/]+)/embed_widget$', 'map_embed_widget', name='map_embed_widget'),
     url(r'^(?P<mapid>[^/]+)/history$', 'ajax_snapshot_history'),
     url(r'^(?P<mapid>\d+)/thumbnail$', 'map_thumbnail', name='map_thumbnail'),
     url(r'^(?P<mapid>[^/]+)/(?P<snapshot>[A-Za-z0-9_\-]+)/view$', 'map_view'),
