+++ geonode/maps/views.py
@@ -28,7 +28,7 @@ from django.core.urlresolvers import reverse
 from django.core.serializers.json import DjangoJSONEncoder
 from django.http import HttpResponse, HttpResponseRedirect, HttpResponseNotAllowed, HttpResponseServerError
 from django.http.response import HttpResponseBadRequest
-from django.shortcuts import render_to_response, get_object_or_404
+from django.shortcuts import render_to_response, get_object_or_404, render
 from django.conf import settings
 from django.template import RequestContext
 from django.utils.translation import ugettext as _
@@ -354,12 +354,36 @@ def map_embed(
     }))
 
 
-def map_embed_widget(request):
+def map_embed_widget(request, mapid,
+                     template='leaflet_maps/map_embed_widget.html'):
+    """Display code snippet for embedding widget.
+
+    :param request: The request from the frontend.
+    :type request: HttpRequest
+
+    :param mapid: The id of the map.
+    :type mapid: String
+
+    :return: formatted code.
     """
-    This function serves as geoserver compatibility.
-    We might need to embed widget on geoserver in the future.
-    """
-    return HttpResponseBadRequest('Sorry, Embed Widget has not supported yet')
+
+    map_obj = _resolve_map(request,
+                           mapid,
+                           'base.view_resourcebase',
+                           _PERMISSION_MSG_VIEW)
+    map_layers = MapLayer.objects.filter(
+        map_id=mapid).order_by('stack_order')
+    layers = []
+    for layer in map_layers:
+        if layer.group != 'background':
+            layers.append(layer)
+
+    context = {
+        'resource': map_obj,
+        'map_layers': layers
+    }
+    message = render(request, template, context)
+    return HttpResponse(message)
 
 
 # MAPS VIEWER #
